FIS-412: Create reusable utils for cta
