+++ src/hztl-foundation/src/lib/utils/cta-utils.ts
@@ -1,0 +1,14 @@
+import { CtaVariants } from './style-param-utils/modules/ctas';
+
+interface CtaStyle {
+  ctaVariant?: CtaVariants;
+}
+
+const getCtaStyle = (ctaStyle: CtaStyle = {}, defaultVariant: CtaVariants) => {
+  return {
+    ...ctaStyle,
+    ctaVariant: ctaStyle?.ctaVariant ?? defaultVariant,
+  };
+};
+
+export { getCtaStyle };
